# server/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      data: {}
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
  api.Response:
    properties:
      data: {}
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
...
  types.TokenData:
    properties:
      token:
        type: string
      user: {}
    type: object
  types.UserInfoResp:
    properties:
      avatar:
        description: 用户头像
        type: string
      background_image:
        description: 用户个人页顶部大图
        type: string
      favorite_count:
        description: 喜欢数
        type: integer
      follow_count:
        description: 关注总数
        type: integer
      follower_count:
        description: 粉丝总数
        type: integer
      id:
        description: 用户 id
        type: integer
      is_follow:
        description: true-已关注，false-未关注
        type: boolean
      name:
        description: 用户名称
        type: string
      work_count:
        description: 作品数
        type: integer
    type: object
...
host: localhost:8811
info:
  contact:
    url: https://github.com/swaggo/swag/blob/master/README_zh-CN.md
  description: This is cxy api docs.
  license:
    name: Apache 2.0
  title: 视频 API
  version: "1.0"
paths:
...
  /user/action:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 通过表单提交进行关注/取关
      operationId: UserAction
      parameters:
      - description: 对方用户id
        in: formData
        name: user_id
        required: true
        type: integer
      - description: 1为关注，2取关
        in: formData
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 用户关注/取关
  /user/follow:
    get:
      description: 通过userId查询用户关注列表
      operationId: UserFollow
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 关注列表
  /user/follower:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 通过userId查询用户粉丝列表
      operationId: UserFollower
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 粉丝列表
  /user/info:
    get:
      consumes:
      - application/json
      description: 通过对方id和我的token获取对方用户信息
      operationId: UserInfo
      parameters:
      - description: 对方用户ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 用户信息
  /user/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 通过表单提交进行用户登录
      operationId: UserLogin
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TokenData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 用户登录
  /user/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 通过表单提交创建用户
      operationId: UserRegister
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 创建用户
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"